Fixes for no-exposed volume support

+++ modules/gigwa/hooks/uninstall.pl
@@ -7,11 +7,13 @@ use Genoring;
 
 ++$|; #no buffering
 
-# Remove Gigwa config.
-my $output = qx(
-  $Genoring::DOCKER_COMMAND run --rm -v $Genoring::VOLUMES_DIR:/genoring -w / alpine rm -rf /genoring/gigwa
-);
-HandleShellExecutionError();
+if (!$ENV{'GENORING_NO_EXPOSED_VOLUMES'}) {
+  # Remove Gigwa config.
+  my $output = qx(
+    $Genoring::DOCKER_COMMAND run --rm -v $Genoring::VOLUMES_DIR:/genoring -w / alpine rm -rf /genoring/gigwa
+  );
+  HandleShellExecutionError();
+}
 
 # Returns 1 when called by "require".
 1;
